requirements.txt
# Graphics2 - NPR渲染系统依赖包
# ========================================
# 基于几何指导的手绘风格三维场景NPR渲染系统

# 核心依赖
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
numba>=0.55.0

# 可选依赖（用于扩展功能）
# scikit-image>=0.19.0  # 图像处理增强功能
# matplotlib>=3.5.0     # 数据可视化和调试

# 开发依赖（可选）
# pytest>=7.0.0         # 单元测试
# black>=22.0.0         # 代码格式化
# flake8>=4.0.0         # 代码检查

# 说明：
# - tkinter 是 Python 标准库，无需额外安装
# - typing 是 Python 3.5+ 标准库
# - json, os, sys 都是标准库
